refactor: RestaurantReviewQueryResponseDao 생성 (#415)
