+++ backend/src/main/java/com/celuveat/restaurant/query/RestaurantReviewQueryService.java
@@ -1,9 +1,7 @@
 package com.celuveat.restaurant.query;
 
-import com.celuveat.restaurant.command.domain.review.RestaurantReview;
-import com.celuveat.restaurant.command.domain.review.RestaurantReviewRepository;
+import com.celuveat.restaurant.query.dao.RestaurantReviewQueryResponseDao;
 import com.celuveat.restaurant.query.dto.RestaurantReviewQueryResponse;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,11 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class RestaurantReviewQueryService {
 
-    private final RestaurantReviewRepository restaurantReviewRepository;
+    private final RestaurantReviewQueryResponseDao restaurantReviewQueryResponseDao;
 
     public RestaurantReviewQueryResponse findAllByRestaurantId(Long restaurantId) {
-        List<RestaurantReview> reviews =
-                restaurantReviewRepository.findAllByRestaurantIdOrderByCreatedDateDesc(restaurantId);
-        return RestaurantReviewQueryResponse.from(reviews);
+        return restaurantReviewQueryResponseDao.findAllByRestaurantId(restaurantId);
     }
 }
