+++ backend/src/main/java/com/celuveat/restaurant/query/dao/RestaurantDetailResponseDao.java
@@ -3,6 +3,9 @@ package com.celuveat.restaurant.query.dao;
 import com.celuveat.celeb.command.domain.Celeb;
 import com.celuveat.restaurant.command.domain.Restaurant;
 import com.celuveat.restaurant.command.domain.RestaurantImage;
+import com.celuveat.restaurant.query.dao.support.RestaurantImageQueryDaoSupport;
+import com.celuveat.restaurant.query.dao.support.RestaurantLikeQueryDaoSupport;
+import com.celuveat.restaurant.query.dao.support.RestaurantQueryDaoSupport;
 import com.celuveat.restaurant.query.dto.RestaurantDetailResponse;
 import com.celuveat.video.command.domain.Video;
 import com.celuveat.video.query.dao.VideoQueryDaoSupport;
